Move out APIBeatmap
